feat: Storage router, create error.go file
