+++ internal/pkg/service/stream/storage/level/local/diskwriter/network/router/balancer.go
@@ -1,8 +1,6 @@
 package router
 
 import (
-	"net/http"
-
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/mapping/recordctx"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/sink/pipeline"
 )
@@ -12,32 +10,4 @@ import (
 // If no pipeline is ready, the NoPipelineReadyError should be returned.
 type Balancer interface {
 	WriteRecord(c recordctx.Context, pipelines []SlicePipeline) (pipeline.RecordStatus, error)
-}
-
-type NoPipelineError struct{}
-
-type NoPipelineReadyError struct{}
-
-func (e NoPipelineError) StatusCode() int {
-	return http.StatusServiceUnavailable
-}
-
-func (e NoPipelineError) ErrorName() string {
-	return "noPipeline"
-}
-
-func (e NoPipelineError) Error() string {
-	return "no pipeline"
-}
-
-func (e NoPipelineReadyError) StatusCode() int {
-	return http.StatusServiceUnavailable
-}
-
-func (e NoPipelineReadyError) ErrorName() string {
-	return "noPipelineReady"
-}
-
-func (e NoPipelineReadyError) Error() string {
-	return "no pipeline is ready"
 }
